Added some documentation and changed toFloat's param type to float

+++ IEEEFloat.cpp
@@ -6,13 +6,13 @@
 #include <sstream>
 #include "IEEEFloat.h"
 
-std::string IEEEFloat::toFloat(double const &num) {
+std::string IEEEFloat::toFloat(float const &num) {
     std::string signBit = computeSignBit(num);
 
     return "";
 }
 
-std::string IEEEFloat::computeSignBit(double const &num) {
+std::string IEEEFloat::computeSignBit(float const &num) {
     if (num < 0) return "1";
     else return "0";
 }
